refact/front/home.jsx: import logic for sorting trips by date
